Fixed the windows-type endings and a small bug
Initially the src was developed on windows so the every file has a windows-type ending apart from README.md and LICENSE which were written on github directly. This commit fixed that problem. Also the initial commit stops working after the plugin is moved to separate folder. This is because when the script is executed, the current folder is not set to the folder where the script stays. Therefore to import the python modules in the cousin folders, a relative path is generated natively in vim script and this fixes this issue.