user autoload to significantly speedup the plugin loading
Previously the plugin was loaded everytime Vim was started. This includes sourcing all the python files due to the import command used in the plugin wrapper. After moving the part bridging python and vim script to the autoload, the converting wrapper, including imported pythons files, are delayed to be loaded until the function call is made. This greately improves the loading speed of the plugin from something around 100 ms to 0.xx ms. The way that autoload works can be found on the LearnVimTheHardWay web. In particlar, the functions names bearing the magic dynamic loading has its definitionFileName in it, for example the file thesaurusPy2Vim.vim contains the function thesaurusPy2Vim#Thesaurus_Lookup().