Fixed the issue of using the plugin with a vim version which only has python3 support.
Previous version does not support a vim version which only has python3
support. This is because the python statement "python ..." only works with
python2. To make it work with python3, those statements are now changed
to "pythonx ...". Vim will choose the version of python according to
the value of the "pyxversion" variable in the .vimrc file. The default value
(2) is for invoking python2. To change to python3, put "set pyx=3" in .vimrc.